feat(stm32): make esc current sense optional
AM32 computes a current from whatever ADC pin its target defaults to, so one
that names none -- TEKKO32_4IN1_F421 among them -- scales a thermistor as amps
and integrates that into the consumed mAh as well. Both now cross the link as
-1 and reach a GCS as NaN, which ESC_STATUS also now reports for an ESC that
never answered rather than zero.

+++ libs/message.hpp
@@ -301,8 +301,10 @@ struct EscTelemetryMsg {
   uint32_t rpm[kEscTelemetryMotorCount];
   uint32_t electrical_rpm[kEscTelemetryMotorCount];
   uint16_t voltage_centivolts[kEscTelemetryMotorCount];
-  uint16_t current_centiamps[kEscTelemetryMotorCount];
-  uint16_t consumption_mah[kEscTelemetryMotorCount];
+  // -1 where the ESC has no shunt. Signed for that alone -- the readings
+  // themselves are never negative.
+  int16_t current_centiamps[kEscTelemetryMotorCount];
+  int32_t consumption_mah[kEscTelemetryMotorCount];
   int16_t temperature_c[kEscTelemetryMotorCount];
   uint8_t valid_mask;
 } __attribute__((packed));
